+++ src/commands/seeder.ts
@@ -35,6 +35,11 @@ const seeder = async (vscode: any, fs: any, path: any) => {
     return;
   }
 
+  if (name === 'Seeder') {
+    vscode.window.showErrorMessage('The file has not been created!');
+    return;
+  }
+
   filename = '/app/Database/Seeds/' + name + '.php';
 
   body = content.replace('{className}', name);