+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Change Log
 
-All notable changes to the "Codeigniter 4 Spark for VSCode" extension will be documented in this file.
+All notable changes to the "CodeIgniter 4 Spark for VSCode" extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.0] - 2023-09-12
+
+### Added
+
+- Add Filename restrictions
+
+### Fixed
+
+- Fix Make Config command
 
 ## [1.4.0] - 2023-09-03
 
@@ -56,7 +66,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.4.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.1.0...v1.2.0